♻️ refactor: add index.tsx file to the @shared folder
