+++ src/components/ButtonCreate.tsx
@@ -2,7 +2,7 @@ import { useRecoilValue, useSetRecoilState } from "recoil";
 import { setCookie } from "cookies-next";
 import { Box } from "@mui/material";
 
-import Button from "./@shared/Button";
+import { Button } from "./@shared";
 import { getOrgs } from "../lib/api";
 import useModal from "../lib/hooks/useModal";
 import loginState from "../lib/recoil/auth";
